requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
evagram==0.0.0
psycopg2==2.9.9
pycodestyle>=2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools>=59.4.0
sqlparse==0.4.4
tzdata==2023.4
